# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MenuAI – AI-powered menu translation and allergen platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants/{restaurantId}/menus:
    get:
      operationId: listMenus
      tags: [menus]
      summary: List menus for a restaurant
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of menus
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Menu"
    post:
      operationId: createMenu
      tags: [menus]
      summary: Create a menu
      parameters:
        - name: restaurantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMenuBody"
      responses:
        "201":
          description: Created menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"
  /menus/{id}:
    get:
      operationId: getMenu
      tags: [menus]
      summary: Get a full menu with sections and dishes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Full menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuWithSections"
        "404":
          description: Not found
    patch:
      operationId: updateMenu
      tags: [menus]
      summary: Update a menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMenuBody"
      responses:
        "200":
          description: Updated menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Menu"
    delete:
      operationId: deleteMenu
      tags: [menus]
      summary: Delete a menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /menus/{id}/translate:
    post:
      operationId: translateMenu
      tags: [menus]
      summary: AI-translate all dishes in a menu to selected languages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslateMenuBody"
      responses:
        "200":
          description: Translation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslateMenuResult"
  /menus/{id}/detect-allergens:
    post:
      operationId: detectAllergens
      tags: [menus]
      summary: AI-detect allergens for all dishes in a menu
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectAllergensResult"

  # ── Sections ─────────────────────────────────────────────────
...
  /public/menu/{slug}:
    get:
      operationId: getPublicMenu
      tags: [public]
      summary: Get the public digital menu for a QR code slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: lang
          in: query
          required: false
          schema:
            type: string
            default: "en"
      responses:
        "200":
          description: Public digital menu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicMenu"
        "404":
          description: Not found
  /public/menu/{slug}/view:
    post:
      operationId: recordDishView
      tags: [public]
      summary: Record a dish view event (analytics)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecordDishViewBody"
      responses:
        "204":
          description: Recorded
components:
  schemas:
...
    Menu:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        name:
          type: string
        qrSlug:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, restaurantId, name, qrSlug, status, createdAt, updatedAt]
    MenuWithSections:
      type: object
      properties:
        id:
          type: integer
        restaurantId:
          type: integer
        name:
          type: string
        qrSlug:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionWithDishes"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, restaurantId, name, qrSlug, status, sections, createdAt, updatedAt]
    CreateMenuBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
      required: [name]
    UpdateMenuBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [draft, published, archived]
    TranslateMenuBody:
      type: object
      properties:
        languages:
          type: array
          items:
            type: string
      required: [languages]
    TranslateMenuResult:
      type: object
      properties:
        translatedDishes:
          type: integer
        languages:
          type: array
          items:
            type: string
      required: [translatedDishes, languages]
    DetectAllergensResult:
      type: object
      properties:
        processedDishes:
          type: integer
        flaggedAllergens:
          type: integer
      required: [processedDishes, flaggedAllergens]

    # ── Section ────────────────────────────────────────────────
...
    SectionWithDishes:
      type: object
      properties:
        id:
          type: integer
        menuId:
          type: integer
        name:
          type: string
        sortOrder:
          type: integer
        dishes:
          type: array
          items:
            $ref: "#/components/schemas/DishDetail"
        createdAt:
          type: string
          format: date-time
      required: [id, menuId, name, sortOrder, dishes, createdAt]
...
    DishDetail:
      type: object
      properties:
        id:
          type: integer
        sectionId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: ["number", "null"]
        currency:
          type: string
        imageUrl:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isSpecial:
          type: boolean
        sortOrder:
          type: integer
        allergens:
          type: array
          items:
            $ref: "#/components/schemas/Allergen"
        dietaryLabels:
          type: array
          items:
            type: string
        translations:
          type: array
          items:
            $ref: "#/components/schemas/DishTranslation"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, sectionId, name, currency, isAvailable, isSpecial, sortOrder, allergens, dietaryLabels, translations, createdAt, updatedAt]
...
    Allergen:
      type: object
      properties:
        id:
          type: integer
        dishId:
          type: integer
        allergenType:
          type: string
          enum: [gluten, dairy, nuts, peanuts, eggs, soy, shellfish, fish, sesame, sulphites, celery, mustard, lupin, molluscs]
        isAiSuggested:
          type: boolean
        isConfirmed:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, dishId, allergenType, isAiSuggested, isConfirmed, createdAt]
...
    DishTranslation:
      type: object
      properties:
        id:
          type: integer
        dishId:
          type: integer
        languageCode:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        culinaryExplanation:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, dishId, languageCode, name, createdAt]

    # ── Analytics ──────────────────────────────────────────────
...
    PublicMenu:
      type: object
      properties:
        restaurantName:
          type: string
        restaurantLogo:
          type: ["string", "null"]
        menuName:
          type: string
        language:
          type: string
        availableLanguages:
          type: array
          items:
            type: string
        sections:
          type: array
          items:
            $ref: "#/components/schemas/SectionWithDishes"
        specials:
          type: array
          items:
            $ref: "#/components/schemas/DishDetail"
      required: [restaurantName, menuName, language, availableLanguages, sections, specials]
    RecordDishViewBody:
      type: object
      properties:
        dishId:
          type: integer
      required: [dishId]